fixed bug page crashes when removing user

+++ src/views/TableView.jsx
@@ -1,10 +1,10 @@
-import { useEffect } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
-import { useNavigate, useParams } from 'react-router-dom'
-import { getTable } from '../store/actions/table.action'
-import emptyChair from '../assets/imgs/empty-chair.svg'
-import { TablePreview } from '../cmps/TablePreview'
-import { tableService } from '../services/table.service'
+import { useEffect } from "react"
+import { useDispatch, useSelector } from "react-redux"
+import { useNavigate, useParams } from "react-router-dom"
+import { getTable } from "../store/actions/table.action"
+import emptyChair from "../assets/imgs/empty-chair.svg"
+import { TablePreview } from "../cmps/TablePreview"
+import { tableService } from "../services/table.service"
 
 export const TableView = () => {
   const table = useSelector((state) => state.tableModule.table)
@@ -14,39 +14,45 @@ export const TableView = () => {
   const params = useParams()
 
   useEffect(() => {
-    dispatch(getTable(params.tableId))
+    ;(async () => {
+      try {
+        await dispatch(getTable(params.tableId))
+      } catch (err) {
+        navigate("/")
+      }
+    })()
   }, [dispatch, params.tableId])
 
   useEffect(() => {
-    if (!table) {
-      navigate('/')
-    } else if (
-      !user ||
-      !table?.users?.find((userInTable) => userInTable.id === user.id)
+    if (
+      table &&
+      (!user ||
+        !table?.users?.find((userInTable) => userInTable.id === user.id))
     ) {
+      console.log("Bar")
       tableService.deleteTableFromStorage()
-      dispatch({ type: 'CLEAR_TABLE', action: null })
-      // navigate('/')
+      dispatch({ type: "SET_TABLE", table: null })
+      navigate("/")
     }
   }, [table, navigate, dispatch, user])
 
   if (!table) return
   let tableParticipants = 4
   return (
-    <div className="table-view">
-      <h1 className="table-number">Table Number</h1>
-      <div className="table">
+    <div className='table-view'>
+      <h1 className='table-number'>Table Number</h1>
+      <div className='table'>
         <h1>{table.tableNumber}</h1>
         {[...Array(tableParticipants)].map((participant, i) => (
           <div className={`chair _${i + 1}`} key={`chair _${i + 1}`}>
             {table.users[i] && (
               <img
                 src={table.users[i].imgUrl}
-                alt="Profile"
-                className="profile-img"
+                alt='Profile'
+                className='profile-img'
               />
             )}
-            <img src={emptyChair} className="empty-chair" alt="Empty Chair" />
+            <img src={emptyChair} className='empty-chair' alt='Empty Chair' />
           </div>
         ))}
       </div>
